Disable stack protector in OpenBSD
This option is enabled by default in OpenBSD and it
generates problems with our crts because it expects
some symbols to be defined. We don't need the stack
protector because this is only a first step to test
the library with gcc and being sure that it is
correct before testing it with scc.