[libmach] Simplify objread() interface
There were two additional funcitons, objsyms() and
objsect() because objread() only read the target
file in a target structure. This mean that after every
objread() we had to a objsyms() and in some cases
a objsect(). This patch simplifies the interface and
it makes objread() to generate the symbols and sections
machine independent form always.