Add distclean target
This target is needed in include/scc/scc because we have files there
that are not removed by clean. These files can be customized by the
user, so it is not a good idea to remove them with clean. The nested
makefile had a distclean target that was never used because the top
makefiles didn't implemented it.